+++ convertkit/admin/setup-wizard/class-convertkit-admin-setup-wizard-restrict-content.php
@@ -15,15 +15,6 @@
 class ConvertKit_Admin_Setup_Wizard_Restrict_Content extends ConvertKit_Admin_Setup_Wizard {
 
 	/**
-	 * Holds the Post Type to generate Members Content for.
-	 *
-	 * @since   2.1.0
-	 *
-	 * @var     string
-	 */
-	public $post_type = 'page';
-
-	/**
 	 * Holds the type of Member's Content to generate (course|download).
 	 *
 	 * @since   2.1.0
@@ -185,6 +176,13 @@ class ConvertKit_Admin_Setup_Wizard_Restrict_Content extends ConvertKit_Admin_Se
 	 */
 	public function process_form( $step ) {
 
+		// Set and authorize the Post Type before processing data.
+		$this->set_post_type();
+		if ( ! convertkit_user_can_create_published_post_type( $this->post_type ) ) {
+			$this->error = __( 'You are not allowed to create and publish this type of content.', 'convertkit' );
+			return;
+		}
+
 		// Run security checks.
 		if ( ! isset( $_REQUEST['_wpnonce'] ) ) {
 			return;
@@ -256,26 +254,10 @@ class ConvertKit_Admin_Setup_Wizard_Restrict_Content extends ConvertKit_Admin_Se
 			wp_die( esc_html__( 'Connect your ConvertKit account in the ConvertKit Plugin\'s settings to get started', 'convertkit' ) );
 		}
 
-		// Get the Post Type.
-		if ( filter_has_var( INPUT_GET, 'ck_post_type' ) ) {
-			$this->post_type = filter_input( INPUT_GET, 'ck_post_type', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
-		} else {
-			$this->post_type = 'page';
-		}
-
-		// Bail if the Post Type isn't supported.
-		if ( ! in_array( $this->post_type, convertkit_get_supported_post_types(), true ) ) {
-			wp_die(
-				sprintf(
-					/* translators: Post Type */
-					esc_html__( 'The post type `%s` is not supported for Member Content.', 'convertkit' ),
-					esc_html( $this->post_type )
-				),
-				esc_html__( 'WordPress Error', 'convertkit' ),
-				array(
-					'back_link' => true,
-				)
-			);
+		// Set and authorize the Post Type.
+		$this->set_post_type();
+		if ( ! convertkit_user_can_create_published_post_type( $this->post_type ) ) {
+			wp_die( esc_html__( 'Sorry, you are not allowed to create and publish this type of content.', 'convertkit' ) );
 		}
 
 		// Define Exit URL to take the user back to the WP_List_Table for the Post Type they were viewing.